backend/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0

